Проект поправлен под работу через репозитории и сделана интеграция с БД

+++ src/main/java/ru/practicum/shareit/item/ItemService.java
@@ -8,9 +8,9 @@ interface ItemService {
 
     ItemDto addItem(String userId, ItemDto itemDto);
 
-    ItemDto editItem(String userId, ItemDto itemDto, int itemId);
+    ItemDto editItem(String userId, ItemDto itemDto, long itemId);
 
-    ItemDto showItem(String userId, int itemId);
+    ItemDto showItem(String userId, long itemId);
 
     Collection<ItemDto> showItems(String userId);
