+++ src/main/java/ru/practicum/shareit/user/UserController.java
@@ -19,7 +19,7 @@ public class UserController {
     private final UserService userService;
 
     @GetMapping("{id}")
-    public User getUser(@PathVariable int id) {
+    public User getUser(@PathVariable long id) {
         log.info("GET /users/{}", id);
         return userService.getUser(id);
     }
@@ -37,13 +37,13 @@ public class UserController {
     }
 
     @PatchMapping("{id}")
-    public User updateUser(@RequestBody User user, @PathVariable int id) {
+    public User updateUser(@RequestBody User user, @PathVariable long id) {
         log.info("PATCH /users");
         return userService.updateUser(id, user);
     }
 
     @DeleteMapping("{id}")
-    public void deleteUser(@PathVariable int id) {
+    public void deleteUser(@PathVariable long id) {
         log.info("DELETE /users");
         userService.deleteUser(id);
     }
